test: remove sleeps from sender account manager
